refactor: modal의 비즈니스 로직과 UI 로직 분리
