+++ src/main/frontend/src/hooks/useModal.jsx
@@ -1,0 +1,15 @@
+import { submitResultFoods, submitResultMachines } from "../api/Products";
+
+export function useModal({ machine, food }) {
+  const close = () => {
+    setResult(false);
+  };
+
+  const confirm = () => {
+    machine && submitResultMachines();
+    food && submitResultFoods();
+    setResult(false);
+  };
+
+  return { close, confirm };
+}
